// backend/src/generated/prisma/models/CourseSubstitution.ts
export type CourseSubstitutionModel = runtime.Types.Result.DefaultSelection<Prisma.$CourseSubstitutionPayload>
// ...
export type $CourseSubstitutionPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "CourseSubstitution"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    originalCourseCode: string
    substituteCourseCode: string
    department: string | null
    startYear: number
    endYear: number | null
    description: string | null
  }, ExtArgs["result"]["courseSubstitution"]>
  composites: {}
}